requirements.txt
asgiref==3.11.0
Django==5.2.9
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==12.1.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
